+++ packages/www/src/documentation/components/navigation.ts
@@ -199,6 +199,7 @@ const components: NavigationSection = {
         { path: 'table', title: 'Table' },
         { path: 'tabs', title: 'Tabs' },
         { path: 'time-format', title: 'TimeFormat' },
+        { path: 'tree', title: 'Tree' },
         { path: 'visually-hidden', title: 'VisuallyHidden' },
       ],
       path: 'content',